ensure node stickiness for replicas and immediate evacuation for controllers
1. Why is this change necessary ?

This is required as part of overall openebs volumes' HA functionality.
This enhancement will ensure a openebs volume replica sticks to the
minion node once placed by K8s scheduler. The node stickiness is applicable
during the cases of minion node being not ready or unreachable. In addition,
another change has been done to evacuate a openebs volume controller when
its hosting node i.e. a minion is unreachable or is not ready.

ref - https://github.com/openebs/mayaserver/issues/91
ref - https://github.com/openebs/mayaserver/issues/95
ref - https://github.com/openebs/openebs/issues/184

2. How does this change address the issue ?

- modified the deployment specs during provisioning of
openebs volume replica(s)

- modified the deployment specs during provisioning of
openebs volume controller

3. How to verify this change ?

- Test in a K8s setup (preferably 1.7)

- kube-controller-manager launch config needs to have
  --feature-gates=TaintBasedEvictions=true

- Browse through the referred issues for more details

4. What side effects does this change have ?

- TaintBasedEvictions is an alpha feature in Kubernetes 1.6 & 1.7.
  This needs to be enabled in Kubernetes setup to use this feature.

- There is some additional code which is currently commented. This is
part of overall refactoring. This may be used or removed in near future.

- Absence of unit tests is a concern